+++ backend/internal/db/user.go
@@ -33,19 +33,19 @@ func GetUserById(ctx context.Context, db *gorm.DB, userID string) (*models.User,
 
 // Tries creating given user in DB
 func CreateUser(ctx context.Context, db *gorm.DB, user models.User) (*models.User, error) {
-	err := gorm.G[models.User](db).
-		Create(ctx, &user)
-	if err != nil {
-		log.Printf("ERROR creating user %v in DB: %v", user, err)
-		return nil, err
-	}
-
 	if user.FullName != "" {
 		log.Printf("Defaulting user.Username %v with its fullname %v", user.UserID, user.FullName)
 		user.Username = user.FullName
 	} else if at := strings.Index(user.Email, "@"); at != -1 {
 		log.Printf("Defaulting user.Username %v with Email (%v) Prefix %v since FullName is empty", user.UserID, user.Email, user.Email[:at])
 		user.Username = user.Email[:at]
+	}
+	
+	err := gorm.G[models.User](db).
+		Create(ctx, &user)
+	if err != nil {
+		log.Printf("ERROR creating user %v in DB: %v", user, err)
+		return nil, err
 	}
 
 	log.Printf("Created user with id %v in DB", user.UserID)