+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [1.1.1+3] - 2026-03-04
+
+### Changed
+
+- Refactored ViewModels
+- Made Login/Logout Process more robust
+
+## [1.1.1] - 2025-01-13
+
+### Fixed
+
+- Prevent empty user.Username state in DB when signing up
+
 ## [1.1.0] - 2025-01-10
 
 ### Added